interview-predictor/cloud-function-uploader/requirements.txt
functions-framework==3.5.0
google-cloud-storage==2.18.0
flask==3.0.0
